+++ textworld/envs/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT license.
 
-from textworld.envs.zmachine.jericho import JerichoEnvironment
-from textworld.envs.zmachine.frotz import FrotzEnvironment
-from textworld.envs.glulx.git_glulx_ml import GitGlulxMLEnvironment as GlulxEnvironment
+from textworld.envs.glulx.git_glulx import GitGlulxEnv
+from textworld.envs.zmachine.jericho import JerichoEnv
+from textworld.envs.wrappers.tw_inform7 import TWInform7