+++ textworld/envs/wrappers/__init__.py
@@ -5,3 +5,4 @@
 from textworld.envs.wrappers.viewer import HtmlViewer
 from textworld.envs.wrappers.recorder import Recorder
 from textworld.envs.wrappers.filter import Filter
+from textworld.envs.wrappers.tw_inform7 import TWInform7