+++ textworld/envs/zmachine/tests/test_zork1.py
@@ -33,8 +33,8 @@ def test_losing_game():
             break
 
     print("Done after {} steps. Score {}/{}.".format(game_state.nb_moves, game_state.score, game_state.max_score))
-    assert game_state.has_lost
-    assert not game_state.has_won
+    assert game_state.lost
+    assert not game_state.won
 
 
 @nottest
@@ -61,5 +61,5 @@ def test_winning_game():
             break
 
     print("Done after {} steps. Score {}/{}.".format(game_state.nb_moves, game_state.score, game_state.max_score))
-    assert game_state.has_won
-    assert not game_state.has_lost
+    assert game_state.won
+    assert not game_state.lost