+++ textworld/helpers.py
@@ -3,24 +3,34 @@
 
 
 import os
-from typing import Optional, Tuple
+from typing import Optional, Tuple, List
 
-from textworld.core import Environment, Agent
+from textworld.core import EnvInfos, Environment, Agent
 from textworld.generator.game import Game, GameOptions
 
-from textworld.envs import GlulxEnvironment
-from textworld.envs import JerichoEnvironment
+from textworld.envs import GitGlulxEnv
+from textworld.envs import JerichoEnv
+from textworld.envs import TWInform7
+from textworld.envs.wrappers import Filter
+from textworld.envs.wrappers.tw_inform7 import MissingGameInfosError
 
 from textworld.agents import HumanAgent
 
 from textworld.generator import make_game, compile_game
 
 
-def start(path: str) -> Environment:
+def start(path: str, infos: Optional[EnvInfos] = None,
+          wrappers: List[callable] = []) -> Environment:
     """ Starts a TextWorld environment to play a game.
 
-    Args:
+    Arguments:
         path: Path to the game file.
+        infos:
+            For customizing the information returned by this environment
+            (see
+            :py:class:`textworld.EnvInfos <textworld.core.EnvInfos>`
+            for the list of available information).
+        wrappers: List of wrappers to apply to the environment.
 
     Returns:
         TextWorld environment running the provided game.
@@ -35,37 +45,41 @@ def start(path: str) -> Environment:
     backend = "glulx" if path.endswith(".ulx") else "zmachine"
 
     if backend == "zmachine":
-        env = JerichoEnvironment(path)
+        env = JerichoEnv(infos)
     elif backend == "glulx":
-        env = GlulxEnvironment(path)
+        env = GitGlulxEnv(infos)
     else:
         msg = "Unsupported backend: {}".format(backend)
         raise ValueError(msg)
 
+    if TWInform7.compatible(path):
+        wrappers = [TWInform7] + list(wrappers)
+
+    # Apply all wrappers
+    for wrapper in wrappers:
+        env = wrapper(env)
+
+    env.load(path)
     return env
 
 
 def play(game_file: str, agent: Optional[Agent] = None, max_nb_steps: int = 1000,
-         wrapper: Optional[callable] = None, silent: bool = False) -> None:
+         wrappers: List[callable] = [], silent: bool = False) -> None:
     """ Convenience function to play a text-based game.
 
     Args:
         game_file: Path to the game file.
         agent: Agent that will play the game. Default: HumanAgent(autocompletion=True).
         max_nb_steps: Maximum number of steps allowed. Default: 1000.
-        wrapper: Wrapper to apply to the environment.
+        wrappers: List of wrappers to apply to the environment.
         silent: Do not render anything to screen.
 
     Notes:
         Use script :command:`tw-play` for more options.
     """
-    env = start(game_file)
     agent = agent or HumanAgent()
-
+    env = start(game_file, wrappers=wrappers + agent.wrappers)
     agent.reset(env)
-    if wrapper is not None:
-        env = wrapper(env)
-
     game_state = env.reset()
     if not silent:
         env.render(mode="human")